Remove ClusterHandlerLibrary and move Cluster.cs to HostServicesAPI project. Also make Cluster implement new interface ICluster for singleton use with dependency injection framework
