--- MultiplayerWebFramework/ClassLibrary1/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace ClassLibrary1
-{
-    public class Class1
-    {
-    }
-}
