--- MultiplayerWebFramework/ClusterHandlerLibrary/ClusterHandlerLibrary.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\MWFModelsLibrary\MWFModelsLibrary.csproj" />
-  </ItemGroup>
-
-</Project>
