+++ MultiplayerWebFramework/HostServicesAPI/Controllers/ClusterController.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Net.Http;
 using System.Text.Json;
 using System.Threading.Tasks;
-using ClusterHandlerLibrary;
+using HostServicesAPI.Interfaces;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using MWFModelsLibrary.Enums;
@@ -17,9 +17,11 @@ namespace HostServicesAPI.Controllers
     public class ClusterController : ControllerBase
     {
         private readonly IHttpClientFactory _clientFactory;
-        public ClusterController(IHttpClientFactory clientFactory)
+        private readonly ICluster _gameInstanceCluster;
+        public ClusterController(ICluster gameInstanceCluster, IHttpClientFactory clientFactory)
         {
             _clientFactory = clientFactory;
+            _gameInstanceCluster = gameInstanceCluster;
         }
         [HttpPost]
         public GameInstanceModel SpinUp([FromBody] JsonElement req)
@@ -45,9 +47,7 @@ namespace HostServicesAPI.Controllers
             string reqArgs     = req.GetProperty("Args").GetString();
             // Request will only give us the game to start, the port, and the arguments when starting it. Everything else will be decided by the Host (us)
 
-            // This is BAD! this is only for testing (maybe add a cluster as a singleton in ConfigureServices instead)
-            Cluster cluster = new Cluster();
-            GameInstanceModel newGameInstance = cluster.SpinUp(reqGameCasted, reqPort, reqArgs);
+            GameInstanceModel newGameInstance = _gameInstanceCluster.SpinUp(reqGameCasted, reqPort, reqArgs);
 
             if (newGameInstance == null)
             {
