+++ MultiplayerWebFramework/HostServicesAPI/HostServicesAPI.csproj
@@ -9,7 +9,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\ClusterHandlerLibrary\ClusterHandlerLibrary.csproj" />
     <ProjectReference Include="..\MWFModelsLibrary\MWFModelsLibrary.csproj" />
   </ItemGroup>
