+++ MultiplayerWebFramework/MultiplayerWebFramework.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 16.0.30204.135
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MWFModelsLibrary", "MWFModelsLibrary\MWFModelsLibrary.csproj", "{844F77AD-4458-4BC7-91D1-BBAEB2A3550F}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClusterHandlerLibrary", "ClusterHandlerLibrary\ClusterHandlerLibrary.csproj", "{CA157205-6708-491F-8A80-A936B0E0E2F0}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HostServicesAPI", "HostServicesAPI\HostServicesAPI.csproj", "{6E3234CD-5405-43D7-A600-FF75C71D0D99}"
 EndProject
 Global
@@ -19,10 +17,6 @@ Global
 		{844F77AD-4458-4BC7-91D1-BBAEB2A3550F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{844F77AD-4458-4BC7-91D1-BBAEB2A3550F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{844F77AD-4458-4BC7-91D1-BBAEB2A3550F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CA157205-6708-491F-8A80-A936B0E0E2F0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CA157205-6708-491F-8A80-A936B0E0E2F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CA157205-6708-491F-8A80-A936B0E0E2F0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CA157205-6708-491F-8A80-A936B0E0E2F0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{6E3234CD-5405-43D7-A600-FF75C71D0D99}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6E3234CD-5405-43D7-A600-FF75C71D0D99}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6E3234CD-5405-43D7-A600-FF75C71D0D99}.Release|Any CPU.ActiveCfg = Release|Any CPU
